+++ babel.config.js
@@ -1,10 +1,3 @@
 module.exports = {
-  presets: [
-    [
-      '@babel/preset-env',
-      { useBuiltIns: 'usage', corejs: '3' }
-    ],
-    '@babel/flow',
-    '@babel/typescript'
-  ],
+  presets: [['@babel/preset-env', {useBuiltIns: 'usage', corejs: '3'}], '@babel/flow', '@babel/preset-typescript'],
 };